tool-calling-from-scratch/requirements.txt
google-generativeai
python-dotenv
Pillow
pytest